+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/logging/MethodAnalyticsTests.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using Xunit;
 
-namespace com.csutil.tests {
+namespace com.csutil.integrationTests {
 
     public class MethodAnalyticsTests {
